Reorganized the classes into folders
