+++ Classes/src/User.java
@@ -1,3 +1,4 @@
+package Classes.src;
 public class User {
     private String username;
     private String password;
